requirements.txt
python-dotenv
pyyaml
pygame
pydantic
requests
langchain
langchain-openai
openai